requirements.txt
# requirements.txt

asttokens==2.4.1
attrs==24.1.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
charset-normalizer==3.3.2
contourpy==1.3.0
control==0.10.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
enum34==1.1.10
executing==2.0.1
fastjsonschema==2.20.0
fonttools==4.53.1
idna==3.7
importlib_resources==6.4.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
libusb-package==1.0.26.2
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.26.4
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pygame==2.6.0
Pygments==2.18.0
PyOpenGL==3.1.7
pyparsing==3.1.4
PyQt5==5.15.10
PyQt5-Qt5==5.15.14
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
pyusb==1.2.1
pyvicon_datastream==0.2.3
pyzmq==26.1.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
scipy==1.13.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
